Address feedback from STJ source-gen PR #51300
