+++ src/libraries/Common/src/System/Runtime/CompilerServices/IsExternalInit.cs
@@ -10,7 +10,12 @@ namespace System.Runtime.CompilerServices
     /// This class should not be used by developers in source code.
     /// </summary>
     [EditorBrowsable(EditorBrowsableState.Never)]
-    public static class IsExternalInit
+#if SYSTEM_PRIVATE_CORELIB
+    public
+#else
+    internal
+#endif
+    static class IsExternalInit
     {
     }
 }
