--- src/libraries/System.Text.Json/gen/IsExternalInit.cs
@@ -1,10 +1,0 @@
-﻿// Licensed to the .NET Foundation under one or more agreements.
-// The .NET Foundation licenses this file to you under the MIT license.
-
-namespace System.Runtime.CompilerServices
-{
-    /// <summary>
-    /// Dummy class so C# init-only properties can compile on NetStandard.
-    /// </summary>
-    internal sealed class IsExternalInit { }
-}
