+++ src/libraries/System.Text.Json/gen/Reflection/ReflectionExtensions.cs
@@ -1,6 +1,7 @@
 ﻿// Licensed to the .NET Foundation under one or more agreements.
 // The .NET Foundation licenses this file to you under the MIT license.
 
+using System.Diagnostics;
 using System.Linq;
 using System.Reflection;
 
@@ -21,12 +22,7 @@ namespace System.Text.Json.SourceGeneration.Reflection
         public static bool IsInitOnly(this MethodInfo method)
         {
             MethodInfoWrapper? methodInfoWrapper = method as MethodInfoWrapper;
-
-            if (methodInfoWrapper == null)
-            {
-                throw new ArgumentException("Expected a MethodInfoWrapper instance.", nameof(method));
-            }
-
+            Debug.Assert(methodInfoWrapper != null);
             return methodInfoWrapper.IsInitOnly;
         }
     }
