+++ src/libraries/System.Text.Json/gen/System.Text.Json.SourceGeneration.csproj
@@ -5,6 +5,8 @@
     <Nullable>enable</Nullable>
     <!-- Suppress warning: XML comment has cref attribute that could not be resolved -->
     <NoWarn>CS1574</NoWarn>
+    <EnableDefaultEmbeddedResourceItems>false</EnableDefaultEmbeddedResourceItems>
+    <UsingToolXliff>true</UsingToolXliff>
   </PropertyGroup>
 
   <PropertyGroup>
@@ -19,15 +21,14 @@
   </ItemGroup>
 
   <ItemGroup>
+    <Compile Include="$(CommonPath)System\Runtime\CompilerServices\IsExternalInit.cs" Link="Common\System\Runtime\CompilerServices\IsExternalInit.cs" />
     <Compile Include="..\Common\JsonIgnoreCondition.cs" Link="Common\System\Text\Json\Serialization\JsonIgnoreCondition.cs" />
     <Compile Include="..\Common\JsonNumberHandling.cs" Link="Common\System\Text\Json\Serialization\JsonNumberHandling.cs" />
     <Compile Include="ClassType.cs" />
     <Compile Include="CollectionType.cs" />
-    <Compile Include="IsExternalInit.cs" />
     <Compile Include="JsonSourceGenerator.cs" />
-    <Compile Include="JsonSourceGeneratorHelper.cs" />
-    <Compile Include="JsonSerializableSyntaxReceiver.cs" />
-    <Compile Include="JsonSourceGeneratorHelper.Generate.cs" />
+    <Compile Include="JsonSourceGenerator.Parser.cs" />
+    <Compile Include="JsonSourceGenerator.Emitter.cs" />
     <Compile Include="ObjectConstructionStrategy.cs" />
     <Compile Include="PropertyMetadata.cs" />
     <Compile Include="Reflection\AssemblyWrapper.cs" />
