--- src/libraries/System.Text.Json/tests/System.Text.Json.Tests/Serialization/IsExternalInit.cs
@@ -1,10 +1,0 @@
-﻿// Licensed to the .NET Foundation under one or more agreements.
-// The .NET Foundation licenses this file to you under the MIT license.
-
-namespace System.Runtime.CompilerServices
-{
-    /// <summary>
-    /// Dummy class so C# 'Record' types can compile on NetStandard and NetFx.
-    /// </summary>
-    public sealed class IsExternalInit { }
-}
